+++ nominal/core/__init__.py
@@ -2,6 +2,18 @@ from nominal.core._event_types import EventType, SearchEventOriginType
 from nominal.core._stream.write_stream import DataStream, LogStream, WriteStream
 from nominal.core._utils.api_tools import LinkDict
 from nominal.core._utils.networking import HeaderProvider
+from nominal.core._utils.properties import (
+    NumericPropertyFilter,
+    PropertyValue,
+    between,
+    eq,
+    gt,
+    gte,
+    lt,
+    lte,
+    neq,
+    not_between,
+)
 from nominal.core._utils.query_tools import ArchiveStatusFilter
 from nominal.core.asset import Asset
 from nominal.core.attachment import Attachment
@@ -83,7 +95,9 @@ __all__ = [
     "Marking",
     "Comment",
     "NominalClient",
+    "NumericPropertyFilter",
     "PointCloudMetadata",
+    "PropertyValue",
     "Run",
     "ScanPattern",
     "SearchEventOriginType",
@@ -106,4 +120,12 @@ __all__ = [
     "Workspace",
     "WorkspaceSearchType",
     "WriteStream",
+    "between",
+    "eq",
+    "gt",
+    "gte",
+    "lt",
+    "lte",
+    "neq",
+    "not_between",
 ]