+++ nominal/experimental/migration/migrator/run_migrator.py
@@ -7,6 +7,7 @@ from typing import Iterable, Mapping, Sequence
 
 from nominal.core import NominalClient
 from nominal.core._utils.api_tools import Link, LinkDict, rid_from_instance_or_string
+from nominal.core._utils.properties import PropertyValue
 from nominal.core.asset import Asset
 from nominal.core.attachment import Attachment
 from nominal.core.run import Run
@@ -25,7 +26,7 @@ class RunCopyOptions(ResourceCopyOptions):
     new_start: datetime | IntegralNanosecondsUTC | None = None
     new_end: datetime | IntegralNanosecondsUTC | None = None
     new_description: str | None = None
-    new_properties: Mapping[str, str] | None = None
+    new_properties: Mapping[str, PropertyValue] | None = None
     new_labels: Sequence[str] | None = None
     new_links: Sequence[str | Link | LinkDict] | None = None
     new_attachments: Iterable[Attachment] | Iterable[str] | None = None